[security] 6-digit PIN with crypto RNG + TLS 1.3 infrastructure
Phase 1 — PIN hardening:
- PIN: 4-digit (u16) → 6-digit (u32), 10K → 1M combinations
- RNG: subsec_nanos() → rand::random() (cryptographically secure)
- Lockout: 3 attempts/60s → 5 attempts/300s
- Protocol: PIN_RESPONSE now sends 4 bytes (u32 LE)

Phase 2 — TLS infrastructure:
- New tls.rs: self-signed cert via rcgen, SHA-256 fingerprint for TOFU
- Dependencies: rustls 0.23, tokio-rustls 0.26, rcgen 0.13, sha2 0.10
- 3 TLS tests: acceptor creation, fingerprint format, cert uniqueness

Tests: 99 → 104 (+5: 2 pairing + 3 TLS)

+++ client/app/src/main/cpp/tcp_client.h
@@ -21,7 +21,7 @@ public:
 
     /// Run the handshake: HELLO → PIN → CONFIG → START.
     /// Returns true if streaming is ready.
-    bool handshake(uint16_t pin);
+    bool handshake(uint32_t pin);
 
     const StreamConfig& getStreamConfig() const { return m_config; }
     bool isConnected() const { return m_connected; }
